Moved code into input file
